[特性新增] 新增 Makefile 文件，使其能够更加方便地构建项目.

+++ tests/hdg/test_lexer.cpp
@@ -107,7 +107,7 @@ TEST(test_lexer, _1){
     expected_tokens.emplace_back(hdg::Token::EF);
 
     ASSERT_EQ(actual_tokens.size(), expected_tokens.size());
-    for (int i=0; i<expected_tokens.size(); i++){
+    for (size_t i=0; i<expected_tokens.size(); i++){
         ASSERT_EQ(expected_tokens[i].getType(), actual_tokens[i].getType());
         ASSERT_EQ(expected_tokens[i].getValue(), actual_tokens[i].getValue());
     }
@@ -125,7 +125,7 @@ TEST(test_lexer, _2){
     expected_tokens.emplace_back(hdg::Token::EF);
 
     ASSERT_EQ(actual_tokens.size(), expected_tokens.size());
-    for (int i=0; i<expected_tokens.size(); i++){
+    for (size_t i=0; i<expected_tokens.size(); i++){
         ASSERT_EQ(expected_tokens[i].getType(), actual_tokens[i].getType());
         ASSERT_EQ(expected_tokens[i].getValue(), actual_tokens[i].getValue());
     }
@@ -148,7 +148,7 @@ TEST(test_lexer, _3){
     expected_tokens.emplace_back(hdg::Token::EF);
 
     ASSERT_EQ(actual_tokens.size(), expected_tokens.size());
-    for (int i=0; i<expected_tokens.size(); i++){
+    for (size_t i=0; i<expected_tokens.size(); i++){
         ASSERT_EQ(expected_tokens[i].getType(), actual_tokens[i].getType());
         ASSERT_EQ(expected_tokens[i].getValue(), actual_tokens[i].getValue());
     }
